requirements.txt
asgiref==3.6.0
auth==0.5.3
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-filter==22.1
django-resized==1.0.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
drf-yasg==1.21.4
eventlet==0.33.3
ez-setup==0.9
falcon==3.1.1
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.72.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
greenlet==2.0.1
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mongoengine==0.26.0
openapi-codec==1.3.2
packaging==23.0
Pillow==9.0.0
protobuf==4.21.12
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pymongo==4.3.3
PyMySQL==1.0.2
pyparsing==3.0.9
python-config==0.1.2
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
utils==1.0.1
Werkzeug==2.2.2
